backend/user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "User service for Genealogy Tree Creator"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic[email]>=1.10.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0",
    "motor>=3.1.1",
    "pymongo>=4.3.0",
]

[project.scripts]
start = "uvicorn:run"

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]
